+++ src/main/java/se499/kayaanbackend/Manual_Generate/Note/entity/NoteImage.java
@@ -15,7 +15,7 @@ import java.time.LocalDateTime;
 public class NoteImage {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer imageID;
+    private Long imageID;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "noteID", nullable = false)
